requirements.txt
openai==1.3.8
pyautogui==0.9.53
pillow==10.0.0
pyttsx3==2.90
python-dotenv==1.0.0
